part 4B
Implementing sorting, i have used age and name as sorting parameters, but the best would be visitor id and name.

Furthermore, intering visitors to the ridehistory and waiting list seems to be complicated where same code is being repeated. Would like to use a for loop to make the code more efficient.